Ransomware Talk - Newcastle BSides
The slides from my talk at Newcastle BSides in September 2021. An overview of the talk can be found below based on the problem statement. 

Problem
As of 2020 researchers have estimated the cost of rectifying a ransomware attack to be $1.85 million. What researchers are not saying is that many of these ransomware attacks have exploited weaknesses in company defences, ones that theoretically should be managed by policies, vulnerabilities. 

Enhanced Problem
Ransomware attacks have become synonymous with Advanced Persistent Threat (APT) actors, with the focus of security researchers linking incidents we all read about in the mainstream news to particular ransomware groups such as REvil / Babuk amongst many more, alongside naming and shaming the company involved. This is despite ransomware itself working on the premise of exploiting weaknesses in defences, namely known (and sometimes undisclosed) vulnerabilities. 

What would happen if the conversation around ransomware changed and instead of purely focusing on threat actors, companies and security teams were encouraged to focus on vulnerabilities! 

This talk focuses on dual discussion needed across the security community of focusing on vulnerabilities that are likely to be weaponised, because let’s face it, threat actors will exploit any weakness they can find. (Thank you Shodan!)